move concurrency utils to _core
